start trying to snip links between kube-apiserver and openshift-apiserver
