+++ pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/customresourcevalidationregistration/cr_validation_registration.go
@@ -3,17 +3,17 @@ package customresourcevalidationregistration
 import (
 	"k8s.io/apiserver/pkg/admission"
 
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/authentication"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/clusterresourcequota"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/config"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/console"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/features"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/image"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/oauth"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/project"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/rolebindingrestriction"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/scheduler"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation/securitycontextconstraints"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/authentication"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/clusterresourcequota"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/config"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/console"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/features"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/image"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/oauth"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/project"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/rolebindingrestriction"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/scheduler"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/securitycontextconstraints"
 )
 
 // AllCustomResourceValidators are the names of all custom resource validators that should be registered
