+++ pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/features/validate_features.go
@@ -13,7 +13,7 @@ import (
 	"k8s.io/apiserver/pkg/admission"
 
 	configv1 "github.com/openshift/api/config/v1"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation"
 )
 
 const PluginName = "config.openshift.io/ValidateFeatureGate"
