+++ pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/project/validate_project.go
@@ -1,4 +1,4 @@
-package scheduler
+package project
 
 import (
 	"fmt"
@@ -7,94 +7,99 @@ import (
 	"k8s.io/apimachinery/pkg/api/validation"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	validationutil "k8s.io/apimachinery/pkg/util/validation"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/apiserver/pkg/admission"
 
 	configv1 "github.com/openshift/api/config/v1"
-	"github.com/openshift/origin/pkg/admission/customresourcevalidation"
+	"github.com/openshift/origin/pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation"
 )
 
-const PluginName = "config.openshift.io/ValidateScheduler"
+const PluginName = "config.openshift.io/ValidateProject"
 
 // Register registers a plugin
 func Register(plugins *admission.Plugins) {
 	plugins.Register(PluginName, func(config io.Reader) (admission.Interface, error) {
 		return customresourcevalidation.NewValidator(
 			map[schema.GroupResource]bool{
-				configv1.Resource("schedulers"): true,
+				configv1.Resource("projects"): true,
 			},
 			map[schema.GroupVersionKind]customresourcevalidation.ObjectValidator{
-				configv1.GroupVersion.WithKind("Scheduler"): schedulerV1{},
+				configv1.GroupVersion.WithKind("Project"): projectV1{},
 			})
 	})
 }
 
-func toSchedulerV1(uncastObj runtime.Object) (*configv1.Scheduler, field.ErrorList) {
+func toProjectV1(uncastObj runtime.Object) (*configv1.Project, field.ErrorList) {
 	if uncastObj == nil {
 		return nil, nil
 	}
 
 	allErrs := field.ErrorList{}
 
-	obj, ok := uncastObj.(*configv1.Scheduler)
+	obj, ok := uncastObj.(*configv1.Project)
 	if !ok {
 		return nil, append(allErrs,
-			field.NotSupported(field.NewPath("kind"), fmt.Sprintf("%T", uncastObj), []string{"Scheduler"}),
+			field.NotSupported(field.NewPath("kind"), fmt.Sprintf("%T", uncastObj), []string{"Project"}),
 			field.NotSupported(field.NewPath("apiVersion"), fmt.Sprintf("%T", uncastObj), []string{"config.openshift.io/v1"}))
 	}
 
 	return obj, nil
 }
 
-type schedulerV1 struct {
+type projectV1 struct {
 }
 
-func validateSchedulerSpec(spec configv1.SchedulerSpec) field.ErrorList {
+func validateProjectSpec(spec configv1.ProjectSpec) field.ErrorList {
 	allErrs := field.ErrorList{}
 
-	if name := spec.Policy.Name; len(name) > 0 {
-		for _, msg := range validation.NameIsDNSSubdomain(spec.Policy.Name, false) {
-			allErrs = append(allErrs, field.Invalid(field.NewPath("spec.Policy.name"), name, msg))
+	if len(spec.ProjectRequestMessage) > 4096 {
+		allErrs = append(allErrs, field.Invalid(field.NewPath("spec.projectRequestMessage"), spec, validationutil.MaxLenError(4096)))
+	}
+
+	if name := spec.ProjectRequestTemplate.Name; len(name) > 0 {
+		for _, msg := range validation.NameIsDNSSubdomain(spec.ProjectRequestTemplate.Name, false) {
+			allErrs = append(allErrs, field.Invalid(field.NewPath("spec.projectRequestTemplate.name"), name, msg))
 		}
 	}
 
 	return allErrs
 }
 
-func (schedulerV1) ValidateCreate(uncastObj runtime.Object) field.ErrorList {
-	obj, allErrs := toSchedulerV1(uncastObj)
+func (projectV1) ValidateCreate(uncastObj runtime.Object) field.ErrorList {
+	obj, allErrs := toProjectV1(uncastObj)
 	if len(allErrs) > 0 {
 		return allErrs
 	}
 
 	allErrs = append(allErrs, validation.ValidateObjectMeta(&obj.ObjectMeta, false, customresourcevalidation.RequireNameCluster, field.NewPath("metadata"))...)
-	allErrs = append(allErrs, validateSchedulerSpec(obj.Spec)...)
+	allErrs = append(allErrs, validateProjectSpec(obj.Spec)...)
 
 	return allErrs
 }
 
-func (schedulerV1) ValidateUpdate(uncastObj runtime.Object, uncastOldObj runtime.Object) field.ErrorList {
-	obj, allErrs := toSchedulerV1(uncastObj)
+func (projectV1) ValidateUpdate(uncastObj runtime.Object, uncastOldObj runtime.Object) field.ErrorList {
+	obj, allErrs := toProjectV1(uncastObj)
 	if len(allErrs) > 0 {
 		return allErrs
 	}
-	oldObj, allErrs := toSchedulerV1(uncastOldObj)
+	oldObj, allErrs := toProjectV1(uncastOldObj)
 	if len(allErrs) > 0 {
 		return allErrs
 	}
 
 	allErrs = append(allErrs, validation.ValidateObjectMetaUpdate(&obj.ObjectMeta, &oldObj.ObjectMeta, field.NewPath("metadata"))...)
-	allErrs = append(allErrs, validateSchedulerSpec(obj.Spec)...)
+	allErrs = append(allErrs, validateProjectSpec(obj.Spec)...)
 
 	return allErrs
 }
 
-func (schedulerV1) ValidateStatusUpdate(uncastObj runtime.Object, uncastOldObj runtime.Object) field.ErrorList {
-	obj, errs := toSchedulerV1(uncastObj)
+func (projectV1) ValidateStatusUpdate(uncastObj runtime.Object, uncastOldObj runtime.Object) field.ErrorList {
+	obj, errs := toProjectV1(uncastObj)
 	if len(errs) > 0 {
 		return errs
 	}
-	oldObj, errs := toSchedulerV1(uncastOldObj)
+	oldObj, errs := toProjectV1(uncastOldObj)
 	if len(errs) > 0 {
 		return errs
 	}
