+++ pkg/cmd/openshift-kube-apiserver/admission/customresourcevalidation/securitycontextconstraints/defaulting_scc.go
