+++ pkg/cmd/openshift-kube-apiserver/admission/namespaceconditions/labelcondition.go
