+++ pkg/cmd/openshift-kube-apiserver/admission/namespaceconditions/namecondition.go
